changed type of id from uint to uint64

+++ db/models/chat.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Chat struct {
-	ID        uint      `gorm:"column:id"`
+	ID        uint64    `gorm:"column:id"`
 	CreatedAt time.Time `gorm:"column:created_at"`
 	IsDead    bool      `gorm:"column:is_dead;default:false"`
 }
